Fix unit tests which fail related to log4j 2.17.2 update
