+++ activemq-unit-tests/src/test/resources/log4j2-test.properties
@@ -19,7 +19,6 @@
 # The logging properties used during tests
 #
 rootLogger.level=INFO
-
 rootLogger.appenderRef.console.ref=Console
 rootLogger.appenderRef.logfile.ref=RollingFile
